feat: Introduce SQLite data collector
With this feature results of a benchmark run can be collected in a
SQLite DB for further processing. The `data-collection.sqlite3` file
will be created/modified in the given output directory (by default
`zbdbench_results`).

See `README.md` for further documentation.

This patch implements the data collection for the
`fio_zone_throughput_avg_lat` benchmark. To implement the data
collection for other benchmarks follow the description in `README.md`
and the implementation of the previously mentioned benchmark in
`data_collector/sqlite_data_collector.py`.

Signed-off-by: Dennis Maisenbacher <dennis.maisenbacher@wdc.com>